Fix Iterative Deepening A*

+++ TP1/Models/SearchTree.cs
@@ -6,7 +6,7 @@ namespace TP1.Models
 {
     public abstract class SearchTree
     {
-        public delegate int HeuristicFunction(Node n);
+        public delegate double HeuristicFunction(Node n);
 
         public Node Root { get; private set; }
         protected SearchTree(State root)
